Lazily evaluate named arguments
An interpolation string need not reference every variable
supplied as a named dot-argument of `glue_data()`, so there is no need to
evaluate them eagerly. In particular

    glue("{x}", x = "blah", y = {Sys.sleep(1); "y"})

ought to be as fast as

    glue("{x}", x = "blah")

and

    glue("blah", x = stop("!")) == as_glue("blah")

To do this we use `delayedAssign()` to evaluate the arguments when they
are referenced, rather than eagerly evaluating them before glue
interpolation.