Improve error clarity
I was playing around with coverage tests in a function using `glue()` and was puzzled by this error. I wasn't sure if it was coming from the R parser, or how to deal with it, since the code looks reasonable enough to me:

```r
glue::glue("{# a comment}")
# Error in glue_data(.x = NULL, ..., .sep = .sep, .envir = .envir, .open = .open,  : 
#   Unterminated comment
```

(I guess this error is for simplicity in more general expressions?)

Anyway, I thought it would be more helpful if it was clearer why it failed / how to fix it.